// src/types/types.ts
export interface EmulatorSettings {
  gbModeOverGbc: boolean;
  useBiosRom: boolean;
  overrideCartridgeType: boolean;
  allowMbcBankAccess: boolean;
  colorizeGameboyPalettes: boolean;
  minimalFullscreen: boolean;
  resizeCanvasDirectly: boolean;
  disallowTypedArrays: boolean;
  useDmgBootRom: boolean;
  smoothResizing: boolean;
  enableChannel1: boolean;
  enableChannel2: boolean;
  enableChannel3: boolean;
  enableChannel4: boolean;
}